requirements.txt
torch
torchvision
nibabel
matplotlib
numpy
Pillow
